refactor: Change widget edit page to settings #308
