+++ src/Widgets/BlogArchives/Manage/Widgets/BlogArchivesSettings.cshtml.cs
@@ -4,12 +4,12 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Newtonsoft.Json;
 using System.Threading.Tasks;
 
-namespace BlogCategories.Manage.Widgets
+namespace BlogArchives.Manage.Widgets
 {
-    public class BlogCategoriesEditModel : PageModel
+    public class BlogArchivesSettingsModel : PageModel
     {
         protected readonly IWidgetService widgetService;
-        public BlogCategoriesEditModel(IWidgetService widgetService)
+        public BlogArchivesSettingsModel(IWidgetService widgetService)
         {
             this.widgetService = widgetService;
         }
@@ -22,7 +22,7 @@ namespace BlogCategories.Manage.Widgets
         /// <param name="widgetId"></param>
         public async Task OnGet(int widgetId)
         {
-            var widget = (BlogCategoriesWidget)await widgetService.GetExtensionAsync(widgetId);
+            var widget = (BlogArchivesWidget)await widgetService.GetExtensionAsync(widgetId);
             WidgetJson = JsonConvert.SerializeObject(widget);
         }
 
@@ -30,7 +30,7 @@ namespace BlogCategories.Manage.Widgets
         /// Updates widget properties. 
         /// </summary>
         /// <param name="widget"></param>
-        public async Task<IActionResult> OnPostAsync([FromBody]BlogCategoriesWidget widget)
+        public async Task<IActionResult> OnPostAsync([FromBody]BlogArchivesWidget widget)
         {
             if (ModelState.IsValid)
             {
