+++ src/Widgets/BlogCategories/Manage/Widgets/BlogCategoriesSettings.cshtml
@@ -1,7 +1,7 @@
 ﻿@page
-@model BlogArchives.Manage.Widgets.BlogArchivesEditModel
+@model BlogCategories.Manage.Widgets.BlogCategoriesSettingsModel
 @{
-    ViewData["Title"] = "BlogArchives Widget Settings";
+    ViewData["Title"] = "BlogCategories Widget Settings";
     Layout = "_SettingsLayout";
 }
 
@@ -21,7 +21,7 @@
             }),
             methods: {
                 save() {
-                    axios.post('/widgets/BlogArchivesEdit', this.widget, this.$root.headers)
+                    axios.post('/widgets/BlogCategoriesEdit', this.widget, this.$root.headers)
                         .then(resp => this.$root.onExtSettingsUpdated({
                             title: this.widget.title,
                             widgetId: this.widget.id,
