+++ src/Widgets/BlogTags/Manage/Widgets/BlogTagsSettings.cshtml
@@ -1,19 +1,17 @@
 ﻿@page
-@model RecentBlogPosts.Manage.Widgets.RecentBlogPostsEditModel
+@model BlogTags.Manage.Widgets.BlogTagsSettingsModel
 @{
-    ViewData["Title"] = "RecentBlogPosts Widget Settings";
+    ViewData["Title"] = "BlogTags Widget Settings";
     Layout = "_SettingsLayout";
 }
 
 <edit-widget inline-template>
     <v-form v-model="valid">
         <v-text-field label="Title" v-model="widget.title" v-on:keydown.enter.prevent="save"></v-text-field>
-        <v-text-field label="Number of posts to show"
-                      v-model="widget.numberOfPostsToShow"
-                      :rules="numberOfPostsToShowRules"></v-text-field>
-        <v-checkbox label="Show post author" v-model="widget.showPostAuthor"></v-checkbox>
-        <v-checkbox label="Show post date" v-model="widget.showPostDate"></v-checkbox>
-        <v-checkbox label="Show post excerpt" v-model="widget.showPostExcerpt"></v-checkbox>
+        <v-text-field label="Max number of tags to display"
+                      v-model="widget.maxTagsDisplayed"
+                      :rules="maxTagsDisplayedRules"></v-text-field>
+        <v-checkbox label="Show post counts" v-model="widget.showPostCount"></v-checkbox>
         <v-btn @@click="save" :disabled="!valid">Save</v-btn>
     </v-form>
 </edit-widget>
@@ -24,15 +22,15 @@
             data: () => ({
                 widget: @Html.Raw(Model.WidgetJson),
                 valid: false,
-                numberOfPostsToShowRules: [
+                maxTagsDisplayedRules: [
                     v => !!v || 'Required',
                     v => /^[0-9]+$/.test(v) || 'Integer only',
-                    v => (parseInt(v) >= 1 && parseInt(v) <= 6) || 'Must be between 1 and 6',
+                    v => (parseInt(v) >= 1 && parseInt(v) <= 10000) || 'Must be between 1 and 10000',
                 ],
             }),
             methods: {
                 save() {
-                    axios.post('/widgets/RecentBlogPostsEdit', this.widget, this.$root.headers)
+                    axios.post('/widgets/BlogTagsEdit', this.widget, this.$root.headers)
                         .then(resp => this.$root.onExtSettingsUpdated({
                             title: this.widget.title,
                             widgetId: this.widget.id,
