+++ src/Widgets/RecentBlogPosts/Manage/Widgets/RecentBlogPostsSettings.cshtml.cs
@@ -4,36 +4,25 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Newtonsoft.Json;
 using System.Threading.Tasks;
 
-namespace SocialIcons.Manage.Widgets
+namespace RecentBlogPosts.Manage.Widgets
 {
-    public class SocialIconsEditModel : PageModel
+    public class RecentBlogPostsSettingsModel : PageModel
     {
         protected readonly IWidgetService widgetService;
-        public SocialIconsEditModel(IWidgetService widgetService)
+        public RecentBlogPostsSettingsModel(IWidgetService widgetService)
         {
             this.widgetService = widgetService;
         }
 
         public string WidgetJson { get; set; }
 
-        public class NewUrl
-        {
-            public string Url { get; set; }
-        }
-
         public async Task OnGet(int widgetId)
         {
-            var widget = (SocialIconsWidget)await widgetService.GetExtensionAsync(widgetId);
+            var widget = (RecentBlogPostsWidget)await widgetService.GetExtensionAsync(widgetId);
             WidgetJson = JsonConvert.SerializeObject(widget);
         }
 
-        public IActionResult OnPostAdd([FromBody]NewUrl newUrl)
-        {
-            var socialLink = SocialIconsWidget.GetSocialLink(newUrl.Url);
-            return new JsonResult(socialLink);
-        }
-
-        public async Task<IActionResult> OnPostAsync([FromBody]SocialIconsWidget widget)
+        public async Task<IActionResult> OnPostAsync([FromBody]RecentBlogPostsWidget widget)
         {
             if (ModelState.IsValid)
             {
