+++ src/Widgets/SocialIcons/Manage/Widgets/SocialIconsSettings.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model SocialIcons.Manage.Widgets.SocialIconsEditModel
+@model SocialIcons.Manage.Widgets.SocialIconsSettingsModel
 @{
     ViewData["Title"] = "SocialIcons Widget Settings";
     Layout = "_SettingsLayout";
