+++ src/Widgets/SocialIcons/Manage/Widgets/SocialIconsSettings.cshtml.cs
@@ -4,25 +4,36 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Newtonsoft.Json;
 using System.Threading.Tasks;
 
-namespace BlogTags.Manage.Widgets
+namespace SocialIcons.Manage.Widgets
 {
-    public class BlogTagsEditModel : PageModel
+    public class SocialIconsSettingsModel : PageModel
     {
         protected readonly IWidgetService widgetService;
-        public BlogTagsEditModel(IWidgetService widgetService)
+        public SocialIconsSettingsModel(IWidgetService widgetService)
         {
             this.widgetService = widgetService;
         }
 
         public string WidgetJson { get; set; }
 
+        public class NewUrl
+        {
+            public string Url { get; set; }
+        }
+
         public async Task OnGet(int widgetId)
         {
-            var widget = (BlogTagsWidget)await widgetService.GetExtensionAsync(widgetId);
+            var widget = (SocialIconsWidget)await widgetService.GetExtensionAsync(widgetId);
             WidgetJson = JsonConvert.SerializeObject(widget);
         }
 
-        public async Task<IActionResult> OnPostAsync([FromBody]BlogTagsWidget widget)
+        public IActionResult OnPostAdd([FromBody]NewUrl newUrl)
+        {
+            var socialLink = SocialIconsWidget.GetSocialLink(newUrl.Url);
+            return new JsonResult(socialLink);
+        }
+
+        public async Task<IActionResult> OnPostAsync([FromBody]SocialIconsWidget widget)
         {
             if (ModelState.IsValid)
             {
